bug fixes, lint fixes, moving role picker to own component
