+++ src/components/__tests__/Submissions.test.jsx
@@ -3,7 +3,7 @@ import { configure, mount } from "enzyme";
 
 import Adapter from "enzyme-adapter-react-16";
 
-import Submissions from "../../components/Pages/Submissions";
+import Submissions from "../Pages/Submissions";
 
 configure({ adapter: new Adapter() });
